lib/qwl/pyproject.toml
[project]
name = "qwl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
